+++ src/data/model/OrganizationMembershipGroups.ts
@@ -15,34 +15,25 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { validateUUID } from "@/types/helpers";
-import { type Timestamp, type UUID } from "@/types/stringTypes";
+import { type UUID } from "@/types/stringTypes";
 import { type components } from "@/types/swagger";
 import { type SetRequired } from "type-fest";
 
-export type OrganizationMemberUser = {
-  userId: UUID;
-  userName?: string;
-  userEmail?: string;
-  serviceAccount?: boolean;
-  deploymentKeyAccount?: boolean;
-  dateJoined?: Timestamp;
+export type OrganizationMembershipGroup = {
+  groupId: UUID;
+  groupName: string;
 };
 
-type Members = SetRequired<
+type Memberships = SetRequired<
   components["schemas"]["Organization"],
   "members"
 >["members"];
 
-export function transformOrganizationMemberUserResponse(
-  user: Members[number]["user"],
-): OrganizationMemberUser {
-  return {
-    userId: validateUUID(user?.id),
-    userName: user?.name,
-    userEmail: user?.email,
-    serviceAccount: user?.service_account,
-    deploymentKeyAccount: user?.deployment_key_account,
-    dateJoined: user?.date_joined,
-  };
+export function transformOrganizationMemberGroupsResponse(
+  groups: Memberships[number]["groups"],
+): OrganizationMembershipGroup[] | undefined {
+  return groups?.map((group) => ({
+    groupId: group.id,
+    groupName: group.name,
+  }));
 }
